Move PositionInputData to new Input System
